Memperbaiki tampilan pada history deteksi ke identifikasi ai

+++ backend/models/plantModel.js
@@ -1,16 +1,17 @@
+// models/plantModel.js
 import { db } from '../database/db.js';
 
 export const insertPlant = async (data) => {
-  const { plant_name, scientific_name, description, care_instructions, photo_url, created_by } = data;
+  const { plant_name, scientific_name, description, care_instructions, photo_url, created_by, health_status, disease_name } = data;
 
   const query = `
     INSERT INTO Plant (
-      plant_name, scientific_name, description, care_instructions, photo_url, created_by
+      plant_name, scientific_name, description, care_instructions, photo_url, created_by, health_status, disease_name
     )
-    VALUES (?, ?, ?, ?, ?, ?)
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?)
   `;
 
-  const values = [plant_name, scientific_name, description, care_instructions, photo_url, created_by];
+  const values = [plant_name, scientific_name, description, care_instructions, photo_url, created_by, health_status, disease_name];
 
   try {
     const [result] = await db.query(query, values); // Asumsi Anda menggunakan mysql2
@@ -21,7 +22,7 @@ export const insertPlant = async (data) => {
   }
 };
 
-export const getPlantHistory = async () => {
+export const getPlantHistory = async (userId) => {
   const query = `
     SELECT 
       plant_id,
@@ -29,13 +30,16 @@ export const getPlantHistory = async () => {
       photo_url,
       createdAt,
       description,
-      care_instructions
+      care_instructions,
+      health_status,
+      disease_name
     FROM Plant
+    WHERE created_by = ?
     ORDER BY createdAt DESC
   `;
 
   try {
-    const [rows] = await db.query(query);
+    const [rows] = await db.query(query, [userId]);
     return rows;
   } catch (error) {
     console.error('Error fetching plant history:', error);
@@ -43,10 +47,10 @@ export const getPlantHistory = async () => {
   }
 };
 
-export const deleteAllPlants = async () => {
-  const query = `DELETE FROM Plant`;
+export const deleteAllPlants = async (userId) => {
+  const query = `DELETE FROM Plant WHERE created_by = ?`;
   try {
-    await db.query(query);
+    await db.query(query, [userId]);
   } catch (error) {
     console.error('Error deleting all plants:', error);
     throw error;
